fix: new instance modal error
* refactor: rewrite new instance modal with web components
