+++ src/renderer/templates/NewInstanceModal.pug
@@ -1,0 +1,31 @@
+h1.header New Instance
+.content
+	form.ui.form#form-newInstance
+		.field
+			label Instance Name
+			.ui.input
+				input(type="text" name="instance-name" placeholder="Unnamed Instance")
+		// type of instance
+		.field
+			label Instance Type
+			.ui.fluid.selection.dropdown#dropdown-type
+				//- id == minecraft type
+				input(type="hidden", name="instance-type")
+				.default.text Select Type
+				.menu
+					.item(data-value="vanilla-release") Vanilla Release
+					.item(data-value="vanilla-snapshot") Vanilla Snapshot
+					.item(data-value="vanilla-historical") Vanilla Old Beta / Alpha
+					.item.disabled(data-value="forge") Forge (Comming Soon!)
+		// instance version
+		.field
+			label Minecraft Version
+			.ui.fluid.search.selection.scrolling.dropdown.disabled#dropdown-id
+				input(type="hidden" name="instance-id")
+				.default.text Select Version
+				.menu
+.actions#action-buttons
+	.ui.fluid.buttons
+		button(type="submit")#submit-newInstanceForm.ui.primary.button Create
+		.or
+		button#cancel-button.ui.cancel.button Cancel
