+++ src/renderer/views/layout.pug
@@ -68,7 +68,7 @@ html(lang="en")
 									center
 										button.ui.tiny.cancel.button.inverted.red Continue without logging in
 		//- These are just placeholders that will be replaced with content when modal is loaded
-		.ui.modal#modal-newInstance
+		.ui.modal#modal-newInstance(is="modal-new-instance")
 		.ui.modal#modal-rename(is="instance-rename-modal")
 		.ui.modal#modal-options(is="instance-options-modal")
 		.ui.modal#modal-saves(is="instance-saves-modal")
